+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `PowerStep.sweet_plot` accepts arguments `pre_trig` and `trig`, indicating
   first index of power pulse and power pulse duration. If these arguments are
   provided, position of the pulse is plotted.
+- `PowerStepSet.to_multipactor_test` to create a `MultipactorTest` from a
+  `PowerStepSet`.
+- `MultipactorTest.interactive_sweet_plot` that allows to click the plot, a new
+  figure opens with the associated `PowerStep`. Only works if `MultipactorTest`
+  was created using `PowerStepSet.to_multipactor_test`.
 
 ## [1.9.1] - 2026-03-11